+++ frappe/integrations/doctype/webhook/webhook.py
@@ -49,6 +49,7 @@ class Webhook(Document):
 					if k == w.fieldname:
 						data[w.key] = v
 		r = requests.post(self.request_url, data=json.dumps(data), headers=headers, timeout=5)
+		r.raise_for_status()
 		frappe.logger().debug({"webhook_success":r.text})
 
 def enqueue_webhook(doc, webhook):
@@ -58,5 +59,38 @@ def enqueue_webhook(doc, webhook):
 	except Exception as e:
 		frappe.throw(_("Error in Webhook"), exc=e)
 
-def webhook_request(doc, webhook):
-	frappe.enqueue('frappe.integrations.doctype.webhook.webhook.enqueue_webhook', doc=doc, webhook=webhook)
+def run_webhooks(doc, method):
+	'''Run webhooks for this method'''
+	if frappe.flags.in_import or frappe.flags.in_patch or frappe.flags.in_install:
+		return
+
+	if not getattr(frappe.local, 'webhooks_executed', None):
+		frappe.local.webhooks_executed = []
+
+	if doc.flags.webhooks == None:
+		webhooks = frappe.cache().hget('webhooks', doc.doctype)
+		if webhooks==None:
+			webhooks = frappe.get_all('Webhook',
+				fields=["name", "webhook_docevent"])
+			frappe.cache().hset('webhooks', doc.doctype, webhooks)
+		doc.flags.webhooks = webhooks
+
+	if not doc.flags.webhooks:
+		return
+
+	def _webhook_request(webhook):
+		if not webhook.name in frappe.local.webhooks_executed:
+			frappe.enqueue("frappe.integrations.doctype.webhook.webhook.enqueue_webhook", doc=doc, webhook=webhook)
+			frappe.local.webhooks_executed.append(webhook.name)
+
+	event_list = ["on_update", "after_insert", "on_submit", "on_cancel", "on_trash"]
+
+	if not doc.flags.in_insert:
+		# value change is not applicable in insert
+		event_list.append('on_change')
+		event_list.append('before_update_after_submit')
+
+	for webhook in doc.flags.webhooks:
+		event = method if method in event_list else None
+		if event and webhook.webhook_docevent == event:
+			_webhook_request(webhook)
