+++ frappe/model/document.py
@@ -169,7 +169,7 @@ class Document(BaseDocument):
 			return
 
 		self.flags.email_alerts_executed = []
-		self.flags.webhooks_executed = []
+		setattr(frappe.local, 'webhooks_executed', [])
 
 		if ignore_permissions!=None:
 			self.flags.ignore_permissions = ignore_permissions
@@ -243,7 +243,7 @@ class Document(BaseDocument):
 			return
 
 		self.flags.email_alerts_executed = []
-		self.flags.webhooks_executed = []
+		setattr(frappe.local, 'webhooks_executed', [])
 
 		if ignore_permissions!=None:
 			self.flags.ignore_permissions = ignore_permissions
@@ -673,7 +673,8 @@ class Document(BaseDocument):
 		out = Document.hook(fn)(self, *args, **kwargs)
 
 		self.run_email_alerts(method)
-		self.run_webhooks(method)
+		from frappe.integrations.doctype.webhook.webhook import run_webhooks
+		run_webhooks(self, method)
 
 		return out
 
@@ -725,44 +726,6 @@ class Document(BaseDocument):
 				_evaluate_alert(alert)
 			elif alert.event=='Method' and method == alert.method:
 				_evaluate_alert(alert)
-
-	def run_webhooks(self, method):
-		'''Run webhooks for this method'''
-		if frappe.flags.in_import or frappe.flags.in_patch or frappe.flags.in_install:
-			return
-
-		if self.flags.webhooks_executed==None:
-			self.flags.webhooks_executed = []
-
-		from frappe.integrations.doctype.webhook.webhook import webhook_request
-
-		if self.flags.webhooks == None:
-			webhooks = frappe.cache().hget('webhooks', self.doctype)
-			if webhooks==None:
-				webhooks = frappe.get_all('Webhook',
-					fields=["name", "webhook_docevent"])
-				frappe.cache().hset('webhooks', self.doctype, webhooks)
-			self.flags.webhooks = webhooks
-
-		if not self.flags.webhooks:
-			return
-
-		def _webhook_request(webhook):
-			if not webhook.name in self.flags.webhooks_executed:
-				webhook_request(self, webhook)
-				self.flags.webhooks_executed.append(webhook.name)
-
-		event_list = ["on_update", "after_insert", "on_submit", "on_cancel", "on_trash"]
-
-		if not self.flags.in_insert:
-			# value change is not applicable in insert
-			event_list.append('on_change')
-			event_list.append('before_update_after_submit')
-
-		for webhook in self.flags.webhooks:
-			event = method if method in event_list else None
-			if event and webhook.webhook_docevent == event:
-				_webhook_request(webhook)
 
 	@staticmethod
 	def whitelist(f):
